created modal for confirmation of deletion of post
