+++ package.json
@@ -8,6 +8,7 @@
     "react-debounce-input": "^3.3.0",
     "react-dom": "^17.0.2",
     "react-icons": "^4.6.0",
+    "react-modal": "^3.16.1",
     "react-router-dom": "^6.4.2",
     "react-scripts": "^4.0.3",
     "react-tooltip": "^4.4.2",
